Work toward unit testing
